fix dependency issue
